requirements.txt
apache-airflow
kafka-python
kafka-python-ng
cassandra-driver
spark
pyspark==3.5.1
pandas